+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.0.3] - Apr 18th 2021
+### Added
+  - The protocol's definition
+  - Client's side scripts specified for Windows (`client_windows.py`) and Linux distros (`client_linux.py`).
+### Updated
+  - Simple implementation of client and server on the defined protocol.
+  > To do list: <br>
+  -- Find ways to handle multiple connections (and addressing them) <br>.
+  -- Find ways to edit the interval times <br>
+
 ## [0.0.2] - Apr 16th 2021
 ### Updated
   - Simple client's computer info getting and sending to server on specified time interval, the details on how the dependencies used should be include in the `client.py` script.